[grid.boundaryinfo] added BoundaryInfoConfigs
I was tired of always figuring out the correct intersection so I decided to
put the default_config() of the BoundaryInfos into separate
BoundaryInfoConfigs, each with the same name as the original BoundaryInfo.
Those are not templatized and can be easily carried around and converted
to a BoundaryInfo using the BoundaryInfoProvider once the IntersectionType
is known.