[common.string] allow vectors to be created from scalars
Historically it was possible to create any type of vector of any requested
length from an expression of the form "1" or "[1]" by interpreting this as
the scalar 1 and filling all components of the vector with this value. That
somehow got lost in the re-factoring of string/configuration; this commit
restores the old behavior.

We could argue, though, that we only want vectors of arbitrary length to be
created from "1", not from "[1]", as the latter looks like an explicit wish
for a vector of length 1.