[grid.intersection] deprecate intersectionContains
This method does not do at all what it is supposed to do! The reason is
that intersection.geometry().local(gobal_point) does not do what we
intended it to do. For a 2d intersection this yields a projection of the
2d global point onto the 1d coordinate system of the intersection in such a
way, that points which are clearly away from the intersection (in 2d
coordinates) are within the intersection (in 1d reference coordinates).
Thus using the reference element of the intersection to `checkInside()`
does not yield correct results.

For 2d, one can use 'contains(intersection, point)' instead, which
(geometrically) checks if the point lies on the intersection. For other
dimensions there is no such method implemented at the moment.